Fix major shell ejection issues with MP40
Upon using the "give weapons" cheat to see what PB3 had to offer, once I had made it to the MP40 I found the shell ejection was broken. This was due to the fact that a right port ejection firearm was ejecting shells from the right side, but were falling to the left. This led to the ejection arc on both the vanilla and PB mode MP40's having shell ejection that arced over the firearm, as opposed to away from the right side of the receiver. I also updated the casings from RifleCasingSpawner, to the UAC SMG's MP40CasingSpawnRight, which fits the MP40 much more appropriately considering the MP40 is chambered in 9x19mm Parabellum.  The fixes listed above also adjust the horizontal and vertical spawn offsets to better appear as if originating from the firearm.