Update Deagle.dec
Shifted Shells to right-side ejection due to the Desert Eagle's centered chamber. This is because, as observed by videos I have seen in researching the issue that the shell usually ejects to the side of the shooter's dominant hand, and if shooter is not careful about kick direction management can get hot brass in the face. Simply added the MP40CasingSpawnRight as opposed to the default ejection for added realism, as the sprites depict a right-dominant wielder. This does not effect the UAC default pistol due to it's directed chamber that is made to eject left.