Update Deagle.dec
Shifted casings to right ejection due to the nature of the Desert Eagle's chamber. As seen in many videos, and discussed by owners, if a user isn't careful with the directional management of the kick, the hot brass can be ejected into the wielder's face. This leads me to believe that ejection on a non-directional chamber (as opposed to the forced left ejection the UAC Standard Issue Pistol exhibits) Largely depends on the dominant hand of the wielder. This in light of the sprite's hand position designates the marine as right-dominant shooter. Nonetheless, this change is simply adding a touch of realism, and all comments are welcome. Also this adds shell ejection to the default hipfire "fire" state.