Add ejection port to left MP40, dualies
I changed the sprite for the left MP40 to accommodate the MP40's real-life ejection port which is opposed the carrier bolt and positioned above the magazine insertion point. This leads the change to show an ejection port upon firing the left MP40. The right one's port is off screen due to angling in right hand, no modifications were made to the right gun.