Added real-world acurate ejection to MP40 sprite
The MP40 had no bolt ejection port. Added it and sprited it according to the bolt's setup, or an open bolt group. This means the bolt is open when ready to fire, and once trigger is pulled, throws bolt forward, chambers, fires, and ejects a round, and then returns to open position. Sprited in bolt assembly on left gun only, due to right gun's position to player.