Removed the extra folders
I just realized that I've accidentally added folders in like a dumbass, so I'm removing all of those changes in the extra slot1 and slot 3 to the original folders. Sorry about that! I'm starting to feel bad now... All of these changes should now be in the right folders.
Shotgun ADS sprites adjusted.