Tracers emit light
I was playing a level with dark areas and noticed the tracers themselves do not emit light, and the bright function felt like it was not enough.

This just adds a light to the tracers spawn state.  The color of the light matches the center of the tracer.

I did not notice a performance hit while using this.  However if you do (say a group of persons who have older/slower hardware) it is simple enough to turn this off or even perhaps hide this behind a cvar.