Cyberdemon laser pointer for railgun
The cyberdemon laser pointer has been repurposed (it is its own actor within the railgun.dec at the end of the file).

Does not add any damage or change the mechanics, just comes on while charging a laser shot.  Added as a single line (A_RailAttack) during the charging states for both fire modes of the laser.