Chamber round with PumpSG with Box Mag
This was a feature that was included originally but subsiquently omitted when the Pump shotgun code additions were re-written.  As promised here is the addition of a pumping animation to chamber a round when you reload the Pump SG with box magazine and zero rounds originally in the SG at time of reload.

Only a A_JumpIf added and a couple of states created.  Uses all existing assets.