Overloaded Cell in Railgun
What this is:  When ejecting a cell during reload it becomes overloaded and will beep for about 100-120 tics and explodes (same explosion as a commando beacon).

I have it set as a jump state of 50%.  To even reach that jump state you must have the cvar on for this feature, have 60 cell in reserve, and fire a full 60cell laser blast shot.

I used stock code for the spawner of the cell, so it ejects to the right but pretty close to the player, a hazard for sure.  That can be modified if need be (I found better success with A_SpawnItemEx)

The A_Jump is at 128, if it should be higher or lower let me know.

As a side note I should make this an issue but I will at least state it here first.  I have been finding issues when holding down fire for the larger laser blasts.  The issues are as follows.

-The startsound during the wind up sounds like is on a 1 tic loop which is pretty loud and jarring.  There are four of them during the wind up so I dont know exactly which one is the culprit (likely 3/4).

-Sometimes I am not allowed to make the full laser blast, stopping randomly between 15-30 (most commonly at 20)

-Sometimes even if I fire just 10-20 cell I get the huge laser blast.

That is what I noticed.  I will copy and paste this as an issue later.  I think there are some improvements lined up for the railgun?