Tracer light is now activated by server CVAR
This is just in case someone would be afraid of the performance hit from the tracer light added earlier.  It now is activated by a server CVAR in the visuals menu.

Now there still is a GL light that comes on upon death/puff which has nothing to do with my light, which is looped during travel of the tracer itself.