Fixed Rifle and Shotgun wheel
Fixed a major issue with the weapon special wheel not working on the Shotgun and DMR.
New change: DMR is now able to use the Weapon special wheel with the DMR upgrade, but now trying to chose the UBGL without the upgrade will display a message saying that you don't have an upgrade. An icon was made for this too.