Deagle and Revolver buffs and...
unloads fixed, Minigun overlays when selecting Death Dealer and switching from Death Dealer fixed, Autoshotgun and Deagle quick melee animations.
Note: ASG quick melee and kicking animations for the drum magazine are a wip, so the correct sprites for the cycled drum won't show. I'm aware of this and it will be fixed.