Altermate tracer for minigun
Just like the tracer for the rifle previously written, except now it is for the minigun.  It is for all modes except for plain gatling mode.  The upgraded fire uses sprites that go well above the crosshair mark so even though I wrote the tracer for that it isn't actually visible (I would suggest having all tracers and projectiles hit point of impact about 12 degrees higher than point of aim center as I have written for myself personally in order to force the player to angle the weapon down (and give more space) 12 degrees to compensate, but that is a different PR for a different day.