GZdoom 4.5.0 List Menu Update
A real simple one.  ListMenus have sizing issues after GZdoom 4.5.0 because now a size parameter needs to be made for each and every listmenu.  "Size Clean" would be the most universal command for this so I added it to the three listmenus in menudef.

Now the listmenus should looks like they did prior to the new version of GZdoom.