Fixed melee grab bug and barrel thrower change
The strange melee grab bug when an explosive barrel is thrown when grabbing a barrel with the fists has been fixed.
Change: If you grab or punch a barrel, you will no longer be able to switch weapons until the barrel is drop or thrown. This is intentional.
Some of BarrelThrower code cleaned up to remove some pointless JumpIfInventory DoKick and consildate the frames together. Also applied the no_switch flag on it.
Explosive Barrel grabbed animation smoothed out.