Starting Loadout
I made this originally in August of 2020 but now it has received attention and PB has been well updated since then.  Same code just updated to remove conflicts.  Will delete old PR.

All done with a Cvar and in decorate.  No Zscript or ACS needed (I have made versions which use both and they are not as good).

TheHylianJuggalo is the one who came up with the names of loadouts and the weapons therin.