Update PBRIFLE.dec
Very fast Code update for JMartinez.  In reference to the alternate tracer code starting on line 805, I see nothing wrong with it.  There is a railattack that is called when the CVAR is turned on, and the bullet is fired with no tracer actor called.  When the CVAR is off you get the normal bullet/tracer.