Axe models
updated the axe models and added Animated models for the thrownAxe animations
thrownAxes cannot be toggled mid flight nor are they affected by culling or q3a, meaning if it is thrown with models on, it will stay a model until it lands. same thing goes for the opposite. if its thrown with models off, it will stay a sprite until the end.
this is because the actor normally only exists for a few seconds at most, so adding a toggle or culling just doesn't make much sense.
*i don't think i need to explain why floating, bobbing and spinning do not affect the thrown actor XD*
*side note: there is a bug with blue blood (in general) which makes this variant impossible to see in game without the aid of console commands