Chaingun mode firing sounds changed to match with Gatling mode
Chaingun mode firing sounds changed to match with Gatling, to me it didn't make sense for the Chaingun mode to have different sounds. Guns don't do that irl. I understand rate of fire differences but still. Repurposed an unused sound to make it sound like a single shot and then mixed it with the other Chaingun firing sound.