Juggalo Menu Cleanup Branch
Removed recursive menu options in the 'global settings' (Spawn) menu that led to the same place.

Renamed the Global Settings menu to 'Spawn Settings", as well as changed the description and variables for the monster and weapon spawns. Did not change any CVar values, but rewrote in such a way so there are no 'double negatives'. On now means on and Off  now means off.

Added proper descriptions to the spawn logic rules, and the menu for it is just a single option. (FIGURE OUT DYNAMIC TOOLTIPS TO HAVE A SINGLE DESCRIPTION CHANGE WHEN THE USER CYCLES OPTIONS)

Cleaned up HUD settings submenu to appear much cleaner, made descriptions clear and concise

Removed the 'Festive Hat' option from the menu, (CVar is still present for Popguy - he wants to make it dynamically appear when a user's computer is in December)

Removed the 'weapon settings' submenu, and migrated the options from it to Spawn and Gameplay settings. Respectively, these were

1 - Weapon Disables from 'weapon' settings' are now migrated into spawn settings> Customize weapon spawns
2 Weapon Progression, Plasma Rifle, and Cyro Rifle settings from 'weapon settings' migrated into Gameplay Settings.

Color-coded all subtitles gold, to be more in line with other settings such as the Tilt++ and PB keybind menus

Downsized the 2 grenade settings in 'Gameplay Settings' to be uniform and appear as one section.