Reloading Tweak Fix (not perfect but workable)
Found a reloading issue when the weapon's being dual wielded. it needs every weapon to have its DualWieldToken definition in place in order for it to properly reload.

There may be merge conflicts but this change will make sure that all the dual wield tokens are in place. These should be very minimal changes. I may tweak this further.