DemonTech Reload Code Trimming and Speed Tweak
The DemonTech Rifle has some extra frames that only seem to serve the purpose of delaying the reload time.
It also has some redundant and identical sprites, which are identical to ones being already used in the normal reload animation as seen when the green crystal turns red. Sprites D6T1D0.png to D6T3D0.png in the ReloadGreen folder of the DemonTech Sprites in the file can be deleted, as well as Sprites D4T2E0.png to D4T2K0.png. These are all duplicates that can save additional space in the PB_Staging file.