Revenant Missile Launcher Rework
Done as a request for ThePopeOfDope(Reloading) on Discord
When selected in your equipment section, the Revenant Missile Launcher can be fired while doing any other action with your weapon that is NOT deselecting from it. The hands are removed from the Missile Launcher sprites.
You also have greater control of how it fires as well. You can hold to keep firing, and you won't be forced to shoot in volleys of 4 with each button press. You can hold for any number of shots, and it doesn't stop until you let your finger off the User1 button, or there's no more ammo for this weapon. However, it cannot be used at the same time as the other equipment types.

Every four shots, the ammo count of the revenant missile launcher goes down by 1. This is tracked by an inventory item that counts the number of shots named Done as a request for ThePopeOfDope(Reloading) on Discord
When selected in your equipment section, the Revenant Missile Launcher can be fired while doing any other action with your weapon that is NOT deselecting from it. The hands are removed from the Missile Launcher sprites.
You also have greater control of how it fires as well. You can hold to keep firing, and you won't be forced to shoot in volleys of 4 with each button press. You can hold for any number of shots, and it doesn't stop until you let your finger off the User1 button, or there's no more ammo for this weapon. However, it cannot be used at the same time as the other equipment types.

Every four shots, the ammo count of the revenant missile launcher goes down by 1. This is tracked by an inventory item that counts the number of shots named PB_RevLauncherFireCount.