Requested Dying Monsters Update
Adjusted the heights of some monsters to properly match their sprites. However, this update alone does come with its own issues; As bullets check for headshots by calculating the height ratio of the body, the head hitboxes are compressed until the point that in order to get a headshot, you basically need to shoot at the top half of the forehead for some particularly shorter dying monster actors.