Quick Fix For Shooting Through Unshootables
While this issue does not pop up all too often, there are often parts of a map designed so that you'd have to use a hitscan weapon to hit a button or enemy. This often is done by using an actor that has the SHOOTABLE flag removed.
With projectiles, the player is unable to shoot past these actors, and while there is a destructible decorations update coming in the future, this still stops the player from progressing. In order for the Destroyable Decorations to work in the first place, they need to be shootable. However, in custom wads, if they have actors that are still not shootable and are either not replaced by Project Brutality or replace the Project Brutality decorations regardless (which can also still leave behind non-shootable actors representing the broken versions of those decorations).

This fix allowing the weapon's bullets to shoot through these non-shootable items will fix all the map compatibility issues that rely on hitscan to progress.